fact_checker/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
chromadb==0.4.24
sentence-transformers==2.7.0
nltk==3.8.1
